enhancement: asset row 컴포넌트 분리
